added config section to config file
