+++ src/parsers/ConfigurationParser.py
@@ -7,7 +7,7 @@ import hashlib
 
 from src.utils import parse_value_with_default
 from src.utils import check_linux_permissions
-
+from src.utils import parse_logging_level
 
 class ConfigParser:
     """
@@ -23,7 +23,7 @@ class ConfigParser:
         self.http_configs = {}
         self.ping_configs = {}
         self.api_config = {}
-
+        self.config_config ={}
     def find_any_yml(self):
         """
         Find file with .yml extension
@@ -193,7 +193,6 @@ class ConfigParser:
                                                                default_value=80)
             self.api_config['timeout'] = parse_value_with_default(content=self.file_content['api'], key='timeout',
                                                                   default_value=10)
-
             if 'startup' in self.file_content['api']:
                 self.api_config['startup'] = {}
                 self.api_config['startup']['test'] = parse_value_with_default(
@@ -224,6 +223,25 @@ class ConfigParser:
             logging.debug(msg='api-startup-exit_on_fail ' + str(self.api_config['startup']['exit_on_fail']))
             logging.debug(msg='api-startup-test-retry_after ' + str(self.api_config['startup']['retry_after']))
 
+        except KeyError:
+            logging.error("Config file error / api / KeyError")
+            exit(-2)
+
+    def parse_config(self):
+        """
+        Pase config section
+        :return:
+        """
+        try:
+            if 'config' in self.file_content:
+                self.config_config['wait'] = parse_value_with_default(
+                                            content=self.file_content['config'], key='wait', default_value=10)
+                log_level = parse_value_with_default(
+                    content=self.file_content['config'], key="log_level", default_value="N/A")
+                self.config_config['log_level'] = parse_logging_level(logging_str=log_level)
+
+                self.config_config['log-file'] = parse_value_with_default(
+                    content=self.file_content['config'], key='log_file', default_value="N/A")
         except KeyError:
             logging.error("Config file error / api / KeyError")
             exit(-2)
